# Speed up gemm by reordering the for loops