@cpuhrsch 2019-03-11
This is ready to go after we did a bit more work on investigating why "beta" is set to 0 for all tests.