@zhangguanheng66 2019-03-12
Avoid using memset() function in branch 1 and 3. 
The updated is able to pass the previously failed test: caffe2_onnx_py2_gcc5_ubuntu16_04_test.
The PR is ready to land. @cpuhrsch